Changed database implementation to fix bugs in update and delete.

+++ app/src/main/java/osu/edu/cashout/Fragments/AccountFragment.java
@@ -2,22 +2,17 @@ package osu.edu.cashout.Fragments;
 
 import android.content.Context;
 import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.app.Fragment;
-import android.util.Log;
 import android.util.Patterns;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.Toast;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.EmailAuthProvider;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DataSnapshot;
@@ -25,24 +20,16 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
-import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.DocumentSnapshot;
-import com.google.firebase.firestore.FirebaseFirestore;
+
 
 import osu.edu.cashout.Activities.LoginActivity;
-import osu.edu.cashout.Activities.SignupActivity;
 import osu.edu.cashout.R;
 import osu.edu.cashout.User;
 
-import static android.content.ContentValues.TAG;
 
 public class AccountFragment extends Fragment implements View.OnClickListener{
     private FirebaseAuth mUserAuth;
-    private FirebaseDatabase mFirebaseDatabase;
-    private DatabaseReference mDatabaseReference;
-    private FirebaseFirestore mFirestore;
-    private Button mUpdateAccountButton;
-    private Button mDeleteAccountButton;
+    private DatabaseReference mDbReference;
     private EditText mFirstName;
     private EditText mLastName;
     private EditText mUsername;
@@ -63,8 +50,8 @@ public class AccountFragment extends Fragment implements View.OnClickListener{
         View v = inflater.inflate(R.layout.fragment_account, container, false);
 
         //Find views by ID
-        mUpdateAccountButton = v.findViewById(R.id.update_button);
-        mDeleteAccountButton = v.findViewById(R.id.delete_button);
+        Button mUpdateAccountButton = v.findViewById(R.id.update_button);
+        Button mDeleteAccountButton = v.findViewById(R.id.delete_button);
         mFirstName = v.findViewById(R.id.first_name);
         mLastName = v.findViewById(R.id.last_name);
         mUsername = v.findViewById(R.id.username);
@@ -78,37 +65,31 @@ public class AccountFragment extends Fragment implements View.OnClickListener{
 
         //Set up necessary FireBase components
         mUserAuth = FirebaseAuth.getInstance();
-        mFirebaseDatabase = FirebaseDatabase.getInstance();
-        mDatabaseReference = mFirebaseDatabase.getReference();
-        mFirestore = FirebaseFirestore.getInstance();
         FirebaseUser currentUser = mUserAuth.getCurrentUser();
         if(currentUser != null) {
-            DocumentReference userRef = mFirestore.collection("users").document(currentUser.getUid());
-            userRef.get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
+            //get the reference to the user that is signed-in
+            mDbReference = FirebaseDatabase.getInstance().getReference("users").child(currentUser.getUid());
+
+            //Read info from the database to display to the user
+            mDbReference.addValueEventListener(new ValueEventListener() {
                 @Override
-                public void onSuccess(DocumentSnapshot documentSnapshot) {
-                    curUser = documentSnapshot.toObject(User.class);
-                    mFirstName.setText(curUser.getFirstName());
-                    mLastName.setText(curUser.getLastName());
-                    mUsername.setText(curUser.getUsername());
-                    mEmailField.setText(curUser.getEmail());
+                public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                    //Called initially, and then any time the information in fields change
+                    curUser = dataSnapshot.getValue(User.class);
+                    if (curUser != null){
+                        mFirstName.setText(curUser.getFirstName());
+                        mLastName.setText(curUser.getLastName());
+                        mUsername.setText(curUser.getUsername());
+                        mEmailField.setText(curUser.getEmail());
+                    }
+                }
+
+                @Override
+                public void onCancelled(@NonNull DatabaseError databaseError) {
+                    //Failed to read value
                 }
             });
         }
-
-        //Read info from the database to display to the user
-        mDatabaseReference.addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                //Called initially, and then any time the information in fields change
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError databaseError) {
-                //Failed to read value
-            }
-        });
-
         return v;
     }
 
@@ -125,37 +106,42 @@ public class AccountFragment extends Fragment implements View.OnClickListener{
                     curUser.setFirstName(mFirstName.getText().toString());
                     curUser.setLastName(mLastName.getText().toString());
                     curUser.setUsername(mUsername.getText().toString());
-                    //May want to add some error handling/listeners
-                    mUserAuth.getCurrentUser().updateEmail(curUser.getEmail())
-                            .addOnCompleteListener(new OnCompleteListener<Void>() {
-                                @Override
-                                public void onComplete(@NonNull Task<Void> task) {
-                                    if (task.isSuccessful()) {
-                                        Log.d("TAG", "User email address updated.");
-                                    }
-                                }
-                            });
-                    if(!mPasswordField.getText().toString().isEmpty()) {
-                        mUserAuth.getCurrentUser().updatePassword(mPasswordField.getText().toString());
+                    if(mUserAuth.getCurrentUser() != null) {
+                        //Change the user's email
+                        mUserAuth.getCurrentUser().updateEmail(curUser.getEmail());
+
+                        //Change the user's password
+                        if (!mPasswordField.getText().toString().isEmpty()) {
+                            mUserAuth.getCurrentUser().updatePassword(mPasswordField.getText().toString());
+                        }
                     }
-                    mFirestore.collection("users").document(mUserAuth.getCurrentUser().getUid()).set(curUser);
-                    mUserAuth.getCurrentUser().reauthenticate(EmailAuthProvider
-                            .getCredential(curUser.getEmail(), "password1234"));
+                    else{
+                        Toast.makeText(getActivity(), "Currently unable to update email and/or password, try again later.",
+                                Toast.LENGTH_SHORT).show();
+                    }
+                    //Set the value of the referenced user to the updated user values
+                    mDbReference.setValue(curUser);
                 }
                 break;
             case R.id.delete_button:
-                //This line is null after changing sign-in information
-                String uid = mUserAuth.getCurrentUser().getUid();
-                mUserAuth.getCurrentUser().delete();
-                mFirestore.collection("users").document(uid).delete();
-                //BUG: Updating the email will not delete the user from the firestore, need to see
-                //  how updating email is done or if mUserAuth needs to be updated
+                //Delete the user in all tables as long as getting the current user is not null
+                if(mUserAuth.getCurrentUser() != null) {
+                    mUserAuth.getCurrentUser().delete();
 
-                //Issue is because changing the login credentials requires that we re-sign in the user,
-                //  otherwise currentUser returns null
-                FirebaseAuth.getInstance().signOut();
-                Intent loginIntent = new Intent(mContext, LoginActivity.class);
-                startActivity(loginIntent);
+                    //Delete the instance of the user in the realtime database
+                    mDbReference.removeValue();
+
+                    //Terminate the session associated to the now delered user
+                    FirebaseAuth.getInstance().signOut();
+
+                    Intent loginIntent = new Intent(mContext, LoginActivity.class);
+                    startActivity(loginIntent);
+                }
+                else{
+                    Toast.makeText(getActivity(), "Currently unable to delete your account, try again later.",
+                            Toast.LENGTH_SHORT).show();
+                }
+
                 break;
         }
